Rename layrry-launcher to layrry-core; layrry-cli to layrry-launcher
