+++ integration-test/it-runner/pom.xml
@@ -33,7 +33,7 @@
   <dependencies>
     <dependency>
       <groupId>org.moditect.layrry</groupId>
-      <artifactId>layrry-cli</artifactId>
+      <artifactId>layrry-launcher</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
@@ -56,7 +56,7 @@
                 <arguments>
                   <argument>-classpath</argument>
                   <classpath/>
-                  <argument>org.moditect.layrry.cli.LayrryRunner</argument>
+                  <argument>org.moditect.layrry.launcher.LayrryLauncher</argument>
                   <argument>--layers-config</argument>
                   <argument>src/test/resources/layers.yml</argument>
                   <argument>Alice</argument>
@@ -70,7 +70,7 @@
                 <arguments>
                   <argument>-classpath</argument>
                   <classpath/>
-                  <argument>org.moditect.layrry.cli.LayrryRunner</argument>
+                  <argument>org.moditect.layrry.launcher.LayrryLauncher</argument>
                   <argument>--layers-config</argument>
                   <argument>src/test/resources/layers.toml</argument>
                   <argument>Alice</argument>
