+++ plugin-example/greeter-runner/pom.xml
@@ -33,7 +33,7 @@
   <dependencies>
     <dependency>
       <groupId>org.moditect.layrry</groupId>
-      <artifactId>layrry-cli</artifactId>
+      <artifactId>layrry-launcher</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
@@ -54,7 +54,7 @@
             <arguments>
               <argument>-classpath</argument>
               <classpath/>
-              <argument>org.moditect.layrry.cli.LayrryRunner</argument>
+              <argument>org.moditect.layrry.launcher.LayrryLauncher</argument>
               <gument>--layers-config</gument>
               <argument>src/test/resources/layers.yml</argument>
             </arguments>
